Optimize font rendering with GPU offloading and branchless shader
- Moved UV normalization from GPU to CPU; stored as normalized vec4 in GlyphMetricsBlock.
- Implemented GPU-side Device Pixel Ratio (DPR) scaling for screen-space arrows via FONT_FLAG_APPLY_DPR (4u) and uDevicePixelRatio uniform.
- Achieved branchless texture sampling in the vertex shader for all glyph types (including synthetic backgrounds/underlines) by pre-calculating center UVs.
- Optimized NDC math in the shader using a pre-calculated uViewportScale uniform.
- Refined bit-packing: expanded to 11-bit glyphId and 9-bit unsigned rotation (0-359 range).
- Added setFloat and setVec2 helpers to AbstractShaderProgram.
- Fixed CI build failure related to KDE Neon GPG key retrieval in build-appimage workflow.
- Maintained 28-byte FontInstanceData and 1px underline visual refinements.

+++ src/opengl/OpenGLTypes.h
@@ -89,7 +89,7 @@ struct NODISCARD ColorVert final
 
 struct NODISCARD GlyphMetrics final
 {
-    glm::ivec4 uvRect{}; // uvX, uvY, uvW, uvH
+    glm::vec4 uvRect{}; // normalized uvX, uvY, uvW, uvH
 };
 static_assert(sizeof(GlyphMetrics) == 16);
 
@@ -124,7 +124,7 @@ struct NODISCARD FontInstanceData final
     {
         // glyphId: 11 bits (0-2047)
         // rotation: 9 bits (unsigned, 0-511) - Infomark angles are 0-359.
-        // flags: 6 bits
+        // flags: 6 bits (Italics=1, NamedColor=2, ApplyDPR=4)
         // namedColorIndex: 6 bits
         const uint32_t uGlyphId = static_cast<uint32_t>(glyphId) & 0x7FFu;
         const uint32_t uRotation = static_cast<uint32_t>(static_cast<uint16_t>(rotation)) & 0x1FFu;
